цц

+++ internal/repository/pg/schemes/bot.sql
@@ -8,7 +8,7 @@ CREATE TABLE IF NOT EXISTS bots (
     ch_link       TEXT DEFAULT '',
     group_link_id INT DEFAULT 0,
     lichka        TEXT DEFAULT '',
-    user_creator  INT  DEFAULT 0,
+    user_creator  BIGINT  DEFAULT 0,
     is_disable    INT  DEFAULT 0,
     created_at    TIMESTAMP DEFAULT now(),
     ch_is_skam    INT DEFAULT 0,
